requirements.txt
beautifulsoup4==4.12.3
pdfplumber==0.11.4
nltk==3.8.1
spacy==3.7.6
wandb==0.17.7
fastapi==0.115.0
uvicorn==0.30.6
prometheus-client==0.21.0
prefect==2.20.2
pytest==8.3.3
requests==2.32.3
pyjwt==2.8.0
rouge-score==0.1.2
transformers==4.41.1
peft==0.10.0
torch==2.2.1
trl==0.8.1
bitsandbytes==0.43.0
accelerate==0.29.2
datasets==2.18.0
huggingface-hub==0.22.2
evaluate==0.4.1
mlflow==2.11.3
rouge-score
python-dotenv==1.0.0
rapidfuzz==3.6.1
psutil==5.9.6
evidently==0.4.8
pytest-cov==4.1.0
codecov==2.1.12
